Allow sess-id to be with 0 until it is fully established
The session module is a bit of a mess right now, because we need to
support the bypass case.
Ideally, we should have a raw session with 0,0 session ids and a reserved local
sess id. And when the session is fully created, all the ids are assigned and keys
installed.