session: Bring some sanity to this module
Now both bypass and regular can co-exist. Primarily created a single cheat function
to set the local session id to 0 (for bypass), even though the session is encrypted.

This should also help in scenarios where unencrypted messages are maliciously sent
on encrypted sessions. The session will just not be found, and hence the message rejected.