This is one mega commit
For getting the retransmission table, I needed to have references to sessions
and exchanges stored in the retrans table. This caused me to put lifetimes on
all possible things, or go with RC::new(). But I couldn't find an RC without a
backing heap.

So instead, for now, have used arrays for sessions and exchanges, so that
pointers to them are easier.

Everything is functional right now, but this requires a major round of cleanup.