Session: Remove the concept of child_local_sess_id
This didn't work, as the same session 0 could be used to create multiple
CASE sessions. This caused problems because once the child_local_sess_id
was used, there was no way to regenerate it.

Easier it is to directly reserve a session id from the Session Mgr, which
is enabled in the current structure.