Transport: Move 'network interface' within SessMgr
Everything is functional *before* this commit. After this commit, the commissioning fails.
This is because in this commit we use the peer-address for TX based on the session-id in the exchange.

But multiple sessions with session-id 0 exist, thus causing the reply to not go out properly. The next commit
will replace sess_id in the exchange structure with sess_idx, fixing this.